[ALPS02274235] sensor: fix build err
[Detail] fix build err

[Solution] fix build err

[Feature] [Android Default] Acclerometer

MTK-Commit-Id: 3b313b52f645644e9c9b33f112f41ad0461790d2

CR-Id: ALPS02274235

+++ drivers/misc/mediatek/barometer/bmp280/bmp280.c
@@ -109,8 +109,8 @@ struct bmp_i2c_data {
 	u8 hw_filter;
 	u8 oversampling_p;
 	u8 oversampling_t;
-	u32 last_temp_measurement;
-	u32 temp_measurement_period;
+	unsigned long last_temp_measurement;
+	unsigned long temp_measurement_period;
 	struct bmp280_calibration_data bmp280_cali;
 
 	/* calculated temperature correction coefficient */
